requirements.txt
openai
faiss-cpu
sentence-transformers
numpy
streamlit
python-dotenv
